Modified API (new => from_buffer, create => new), improve docs.
Renamed Element::new to Element::from_buffer, and Element::create to
Element::new. Reordered the functions for "new" to appear first, as
seems to be the convention.

Added an example in the documentation of how to use the XmlWriter and
Element::new.

Fixed the existing example by replacing println! by panic! in case of
errors. Indeed the XML was erroneous, and the parser was emitting
errors, but because we didn't panic, the test was passing.

Fixed the XML to make the test pass.