Add a constructor to XmlDecl.
Previously, the only way for _users_ of the library to construct an `XmlDecl`
was to parse it using an `XmlReader`. This change adds a constructor that 'renders'
an XmlDecl from a _version_, an optional _encoding_ and an optional _standalone_
attribute value (these are the only 3 attributes allowed on XML declarations).