Fix serialization of structs
Attributes were being created as child elements.

This works correctly if the simple types that should be attributes are
kept at the beginning of the struct and the more complex types that
should be children are after all of the simple types.

To solve this correctly, there should probably be two Vecs, one for the
attributes and one for the children.