Remove the ability to use SliceReader with raw bytes.
In the near future, decoding will be performed automatically as the
input is read. If the input has an unknown encoding, it must be decoded
first, necessitating a buffer. Therefore only the buffered
implementation can be used for `Reader::from_bytes()`

If the encoding of the bytes is known up-front, you can decode them
up-front and subsequently use `Reader::from_str()` if desired.