ns: Remove `DerefMut` into Reader because we do not want that user call non-namespace-aware mutation methods
Because `impl DerefMut<Target=Reader> for NsReader` is removed, all Reader's
methods that accept `&mut self` won't be available for the NsReader, including
configuration methods.

Because we need a way to configure our reader, we need to reimplement necessary
methods that accepts `&mut self`. In order to not duplicate the code and the doc
of this configure methods, they was put into the macro `configure_methods!` and
embed to both readers.