Indicate minimum version 2.1 for memchr dependency (#562)
The 2.0 family of memchr doesn't build in this crate.  This is hard to
see using cargo minimal-versions because tokio depends on memchr 2.2,
which raises the dependency using minimal-versions even when tokio is
disabled.  This is a known issue with cargo-minimal-versions:

https://github.com/taiki-e/cargo-minimal-versions/issues/6

So everything looks fine from the perspective of this crate, but
consumers who use quick_xml without the optional features that enable
the tokio dependency have a minimal version down to 2.0, which breaks.

To reproduce in this crate, I commented out the tokio and async_tokio
lines in Cargo.toml and ran cargo minimal-versions, which produces the
errors, and emulates the conditions an external user might experience.