Trim `Text` events after DOCTYPE so spaces does not produce an event.
Otherwise consequent `Text` events (which is possible if their delimited by
Comment or PI events, which is skipped) will be merged but not trimmed.
That will lead to returning a `Text` event when try to call `deserialize_struct`
or `deserialize_map` which will trigger `DeError::ExpectedStart` error.

The incorrect trim behavior was introduced in #581, when DocType event began to be processed