Inline calls of `ser_name` and `serialize_*_variant` in all `serialize_*_variant`s
The code was simply copied with adaptation of parameter names.

This is required, because we want to change representation of internal `enum` fields
to match the current `$text` representation, but keep representation of top-level
enums unchanged (because there are no field for them that could be renamed to `$value`)

That means that

  enum Enum {
    Unit,
  }

would be represented differently when used as a top-level type or as a type of field:

<!-- top-level representation -->
<Unit/>

<!-- in-field representation (in field with name "field") -->
<field>Unit</field>

We need to change default in-field representation to be compatible with serde changes
since >=1.0.181. To get the old representation we need to rename the field to `$value`.
Actually, this is even more correct way, because I noticed, that currently serializer
and deserializer worked differently and a `de` module documentation contains an error,
which was unnoticed, because that code snippet is not tested. That will be fixed in next
commits.