Immediately create ElementSerializers with the correct key
After inlining we could notice that some unnecessary work are done.
This commit slightly changes the behavior, because construction of
XmlName checks the allowed names to be correspond to XML 1.1 rules
for names. This is not a problem, because we do not want to check
names that will not be used in the generated XML.