Regenerate ISO-8859-8-I.xml due to error in its content
CDATA section was formed incorrectly and instead was recognized as a Start tag.

File introduced in PR #465 and was made manually and not using generator, because
WHATWG does not have definition of this encoding as a separate entry in index.json.
Actually, this encoding the same as ISO-8859-8, but influences layout direction
when render text.

Wikipedia:
The WHATWG Encoding Standard used by HTML5 treats ISO-8859-8 and ISO-8859-8-I
as distinct encodings with the same mapping due to influence on the layout direction

So generator was fixed and file regenerated