Fix #580: deserialize newtypes using Visitor::visit_newtype_struct instead of visit_seq
The problem was in `SeqItemDeserializer::deserialize_newtype_struct`. Previously it was
the same as `SeqItemDeserializer::deserialize_seq`, but deserializing sequences in this
deserializer assumes, that those sequences are `xs:list`s, because deserializer itself
represents a list element.

In the original bug report an `UnexpectedEof` is returning. The error was changed in the
pre-previous commit (3d5ed696cd5b8d54298af2d745bec6054314431f) and in this it is fixed.

The following tests checks that all's OK:

- Deserializer::deserialize_newtype_struct is reached by:
  - serde-de:
    - newtype::excess_attribute
    - newtype::simple
  - serde-se:
    - with_root::newtype
    - without_root::newtype

- MapValueDeserializer::deserialize_newtype_struct is reached by:
  - serde-issues:
    - issue343

- SeqItemDeserializer::deserialize_newtype_struct is reached by:
  - serde-issues:
    - issue580