Get rid of deserialize_option! macro
MapValueDeseriaizer and SeqItemDeserializer should emit error on DeEvent::Eof
instead of calling Visitor::visit_none(). We delegate emitting appropriate error
to the Some variant.

The content of deserialize_option! macro was inlined and handling of DeEvent::Eof
was removed from MapValueDeseriaizer and SeqItemDeserializer.

This does not change existing behavior, because Eof in the middle of struct anyway
would be captured by `read_to_end()` call at the end of structure deserialization,
but this change will report an error in appropriate place, which would important
when we implement error position tracking