Move tests of enum with special `$text` and `$value` fields in enum to ContentSerializer tests
In the next commits representation of enums will change and ElementSerializer won't
anymore support serialization of enum struct variants. In order to not lost tests
move them into ContentSerializer. It is also not meaningless (for `$value`), since it is
ContentSerializer which is responsible for serializing `$value` fields and struct
variants will be possible to serialize only in `$value` fields, i. e. they will be
serialized by ContentSerializer. For the `$text` fields tests is moved only to keep them.