Parse `<?xml?>` as XML declaration instead of processing instruction
This is more consistent with the way how we check presence of the mandatory
`version` attribute in XML declaration. That check currently only performed
when you try to read version. Also, technically in the grammar space is
a part of version info: https://www.w3.org/TR/xml11/#NT-XMLDecl

Fixed (3):
    syntax::decl::normal1::async_tokio
    syntax::decl::normal1::borrowed
    syntax::decl::normal1::buffered