Allow `&#0;` character references
We should either restrict all invalid characters both in literal form and
as character references, or none of them. Disallowing only the one character
is inconsistently. Because checking literal forms means that we should decode
and check all the input, this will influence performance. We are not ready to
get that performance lost for now. Users of the Reader API could do their own
checks themselves.