Change Writer errors to only io errors
The other error types in `crate::errors::Error` are not applicable
to the Writer. Only io error can be given and putting the io error
in an `std::sync::Arc` and making users match on all other kinds is
not needed. The returned `std::io::Error` can still be turned into
`crate::errors::Error` if needed, but on demand.